update db via jet

+++ internal/database/proglv/public/model/published_task_codes.go
@@ -9,5 +9,5 @@ package model
 
 type PublishedTaskCodes struct {
 	TaskCode string `sql:"primary_key"`
-	TaskID   int32
+	TaskID   int64
 }
